+++ src/app/[variants]/(main)/agent/_layout/Sidebar/Cron/CronTopicGroup.tsx
@@ -1,9 +1,10 @@
 'use client';
 
-import { Flexbox, Icon, Text } from '@lobehub/ui';
+import { Block, Center, Flexbox, Icon, Text } from '@lobehub/ui';
 import { cssVar } from 'antd-style';
 import { Clock } from 'lucide-react';
 import { type MouseEvent, memo, useCallback } from 'react';
+import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 
 import { useRouter } from '@/app/[variants]/(main)/hooks/useRouter';
@@ -27,6 +28,7 @@ interface CronTopicGroupProps {
 }
 
 const CronTopicGroup = memo<CronTopicGroupProps>(({ cronJob, cronJobId, topics }) => {
+  const { t } = useTranslation('setting');
   const { aid } = useParams<{ aid?: string }>();
   const router = useRouter();
   const handleOpenCronJob = useCallback(
@@ -38,30 +40,34 @@ const CronTopicGroup = memo<CronTopicGroupProps>(({ cronJob, cronJobId, topics }
     [aid, cronJobId, router],
   );
 
-  const cronJobName = cronJob?.name || `Cron Job ${cronJobId.slice(-8)}`;
+  const cronJobName = cronJob?.name || t('agentCronJobs.unnamedTask');
   const isEnabled = cronJob?.enabled ?? false;
 
   return (
-    <Flexbox gap={1}>
-      <Flexbox
+    <Flexbox gap={2}>
+      <Block
         align="center"
-        gap={6}
-        height={24}
+        clickable
+        gap={8}
+        height={32}
         horizontal
         onClick={handleOpenCronJob}
-        paddingInline={8}
-        style={{ cursor: 'pointer', opacity: isEnabled ? 1 : 0.6, overflow: 'hidden' }}
+        paddingInline={4}
+        style={{ opacity: isEnabled ? 1 : 0.5, overflow: 'hidden' }}
+        variant="borderless"
       >
-        <Icon icon={Clock} style={{ color: cssVar.colorTextDescription, opacity: 0.7 }} />
-        <Text ellipsis fontSize={12} style={{ flex: 1 }} type={'secondary'} weight={500}>
+        <Center flex="none" height={24} width={24}>
+          <Icon color={cssVar.colorTextDescription} icon={Clock} size={16} />
+        </Center>
+        <Text ellipsis fontSize={12} style={{ flex: 1 }} type="secondary" weight={500}>
           {cronJobName}
         </Text>
         {topics.length > 0 && (
           <Text style={{ color: cssVar.colorTextDescription, fontSize: 11 }}>{topics.length}</Text>
         )}
-      </Flexbox>
+      </Block>
       {topics.length > 0 && (
-        <Flexbox gap={1} paddingBlock={1}>
+        <Flexbox gap={2} paddingBlock={2}>
           {topics.map((topic) => (
             <CronTopicItem key={topic.id} topic={topic} />
           ))}
